+++ vignettes/references.bib
@@ -1,0 +1,49 @@
+
+@book{posthuma_species_2001,
+	title = {Species sensitivity distributions in ecotoxicology},
+	publisher = {CRC press},
+	author = {Posthuma, Leo and Suter II, Glenn W and Traas, Theo P},
+	year = {2001},
+	url = {https://www.crcpress.com/Species-Sensitivity-Distributions-in-Ecotoxicology/Posthuma-II-Traas/p/book/9781566705783}
+}
+
+@Article{fitdistrplus,
+    title = {{fitdistrplus}: An {R} Package for Fitting Distributions},
+    author = {Marie Laure Delignette-Muller and Christophe Dutang},
+    journal = {Journal of Statistical Software},
+    year = {2015},
+    volume = {64},
+    number = {4},
+    pages = {1--34},
+    url = {http://www.jstatsoft.org/v64/i04/}
+  }
+
+  @Book{ggplot2,
+    author = {Hadley Wickham},
+    title = {ggplot2: Elegant Graphics for Data Analysis},
+    publisher = {Springer-Verlag New York},
+    year = {2016},
+    isbn = {978-3-319-24277-4},
+    url = {http://ggplot2.org}
+  }
+
+ @Manual{r,
+    title = {R: A Language and Environment for Statistical Computing},
+    author = {{R Core Team}},
+    organization = {R Foundation for Statistical Computing},
+    address = {Vienna, Austria},
+    year = {2018},
+    url = {https://www.R-project.org/}
+  }
+
+
+@book{burnham_model_2002,
+	address = {New York},
+	title = {Model {Selection} and {Multimodel} {Inference}: {A} {Practical} {Information}-{Theoretic} {Approach}},
+	volume = {Second Edition},
+	publisher = {Springer},
+	author = {Burnham, Kenneth P. and Anderson, David R.},
+	year = {2002},
+	keywords = {sgpd}
+}
+
