+++ NAMESPACE
@@ -32,6 +32,7 @@ export(GeomXribbon)
 export(StatSsd)
 export(StatSsdcens)
 export(comma_signif)
+export(dgompertz)
 export(dlgumbel)
 export(dllog)
 export(geom_hcintersect)
@@ -42,10 +43,13 @@ export(is.fitdistcens)
 export(is.fitdists)
 export(is.fitdistscens)
 export(npars)
+export(pgompertz)
 export(plgumbel)
 export(pllog)
+export(qgompertz)
 export(qlgumbel)
 export(qllog)
+export(rgompertz)
 export(rlgumbel)
 export(rllog)
 export(ssd_cfplot)

[…]
