+++ NAMESPACE
@@ -35,6 +35,7 @@ export(comma_signif)
 export(dgompertz)
 export(dlgumbel)
 export(dllog)
+export(dpareto)
 export(geom_hcintersect)
 export(geom_ssd)
 export(geom_xribbon)
@@ -46,12 +47,15 @@ export(npars)
 export(pgompertz)
 export(plgumbel)
 export(pllog)
+export(ppareto)
 export(qgompertz)
 export(qlgumbel)
 export(qllog)
+export(qpareto)
 export(rgompertz)
 export(rlgumbel)
 export(rllog)
+export(rpareto)
 export(ssd_cfplot)
 export(ssd_ecd)
 export(ssd_fit_dist)
@@ -65,9 +69,6 @@ import(ggplot2)
 import(purrr)
 import(stats)
 importFrom(FAdist,dllog)
-importFrom(VGAM,dgompertz)
-importFrom(VGAM,gompertz)
-importFrom(VGAM,paretoff)
 importFrom(VGAM,vglm)
 importFrom(fitdistrplus,bootdist)
 importFrom(fitdistrplus,bootdistcens)
