### 2018-11-14 16713b2
remove magrittr
