+++ R/gof.R
@@ -62,7 +62,7 @@ ssd_gof.fitdist <- function(x, ...) {
   bic <- x$bic
 
   if(n >= 8) {
-    x %<>% fitdistrplus::gofstat()
+    x <- fitdistrplus::gofstat(x)
     ad <- x$ad
     ks <- x$ks
     cvm <- x$cvm
@@ -85,13 +85,13 @@ ssd_gof.fitdistcens <- function(x, ...) {
 #' @describeIn ssd_gof Goodness of Fit
 #' @export
 ssd_gof.fitdists <- function(x, ...) {
-  x %<>% map_df(ssd_gof)
+  x <- map_df(x, ssd_gof)
   if(!is.null(x$aicc)) {
     x$delta <- x$aicc - min(x$aicc)
   } else # aicc not defined for censored data
     x$delta <- x$aic - min(x$aic)
   x$weight <- exp(-x$delta/2)/sum(exp(-x$delta/2))
-  x$weight %<>% round(3)
-  x$delta %<>% round(3)
+  x$weight <- round(x$weight, 3)
+  x$delta <- round(x$delta, 3)
   x
 }
