+++ DESCRIPTION
@@ -36,7 +36,6 @@ Imports:
     dplyr,
     FAdist,
     grid,
-    magrittr,
     plyr,
     purrr,
     tibble,
