+++ NAMESPACE
@@ -78,6 +78,4 @@ importFrom(graphics,title)
 importFrom(grid,gpar)
 importFrom(grid,grobName)
 importFrom(grid,polygonGrob)
-importFrom(magrittr,"%<>%")
-importFrom(magrittr,"%>%")
 importFrom(scales,percent)
