+++ README.Rmd
@@ -119,7 +119,7 @@ If you would like to contribute to the package, please see our
 [CONTRIBUTING](CONTRIBUTING.md) guidelines.
 
 Please note that this project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). 
-By participating in this project you agree to abide by its terms.
+By participating in this project you agree to abide by its terms
 
 ## License
 
