+++ DESCRIPTION
@@ -36,14 +36,14 @@ Imports:
     FAdist,
     grid,
     purrr,
-    tibble,
     scales,
     VGAM
 Suggests:
     covr,
     knitr,
     rmarkdown,
-    testthat
+    testthat,
+    tibble
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.1.1
