### 2018-11-14 d33790e
rm purrr
