+++ R/internal.R
@@ -3,3 +3,11 @@ as_conditional_tibble <- function(x) {
     x <- tibble::as_tibble(x)
   x
 }
+
+safely <- function(.f) {
+  function(...) {
+    x <- try(.f(...), silent = TRUE)
+    if(inherits(x, "try-error")) return(list(result = NULL, error = as.character(x)))
+    list(result = x, error = NULL)
+  }
+}
