+++ NAMESPACE
@@ -66,7 +66,6 @@ export(ssd_plot)
 export(stat_ssd)
 import(checkr)
 import(ggplot2)
-import(purrr)
 import(stats)
 importFrom(FAdist,dllog)
 importFrom(VGAM,vglm)
