+++ vignettes/ssdca.R
@@ -63,7 +63,7 @@ fit
 plot(fit)
 
 ## ------------------------------------------------------------------------
-data(fluazinam)
+data(fluazinam, package = "fitdistrplus")
 head(fluazinam)
 
 ## ------------------------------------------------------------------------
