+++ NAMESPACE
@@ -69,8 +69,6 @@ import(ggplot2)
 import(stats)
 importFrom(FAdist,dllog)
 importFrom(VGAM,vglm)
-importFrom(fitdistrplus,bootdist)
-importFrom(fitdistrplus,bootdistcens)
 importFrom(graphics,par)
 importFrom(graphics,plot)
 importFrom(graphics,title)
