+++ NAMESPACE
@@ -69,9 +69,6 @@ import(ggplot2)
 import(stats)
 importFrom(FAdist,dllog)
 importFrom(VGAM,vglm)
-importFrom(graphics,par)
-importFrom(graphics,plot)
-importFrom(graphics,title)
 importFrom(grid,gpar)
 importFrom(grid,grobName)
 importFrom(grid,polygonGrob)
