+++ R/utils.R
@@ -81,7 +81,7 @@ is.fitdistscens <- function(x) {
 #' @param ... Unused.
 #' @export
 #' @examples
-#' nobs(boron_lnorm)
+#' stats::nobs(boron_lnorm)
 nobs.fitdist <- function(object, ...) object$n
 
 #' Number of Observations
@@ -90,12 +90,12 @@ nobs.fitdist <- function(object, ...) object$n
 #' @param ... Unused.
 #' @export
 #' @examples
-#' nobs(boron_lnorm)
+#' stats::nobs(boron_lnorm)
 nobs.fitdistcens <- function(object, ...) nrow(object$censdata)
 
 #' @export
 nobs.fitdists <- function(object, ...) {
-  ns <- vapply(object, nobs, 1L)
+  ns <- vapply(object, stats::nobs, 1L)
   if(!all(ns == ns[1]))
     stop("the fitdists must have the same number of observations", call. = FALSE)
   names(ns) <- NULL
@@ -145,7 +145,7 @@ comma_signif <- function(x, digits = 1, ...) {
 }
 
 ggname <- function(prefix, grob) {
-  grob$name <- grobName(grob, prefix)
+  grob$name <- grid::grobName(grob, prefix)
   grob
 }
 
