### 2018-11-15 37a2443
remove unnecessary ggplot2:: from code
