### 2018-11-15 b2ed5a5
test fit not print dists
