### 2018-11-15 12110ae
not check.attributes test ssd_gof
