### 2018-11-15 #1082 @whedon
[REVIEW]: ssdtools: An R package to fit Species Sensitivity Distributions (SSDs)
**Submitting author:** @joethorley (<a href="http://orcid.org/0000-0002-7683-4592">Joseph Thorley</a>)
**Repository:** <a href="https://github.com/bcgov/ssdtools" target ="_blank">https://github.com/bcgov/ssdtools</a>
**Version:** v0.0.3
