### 2018-11-15 a683281
rename vignette ssdca as ssdtools
