### 2018-11-15 1296a68
update manual
