+++ vignettes/ssdtools.Rmd
@@ -1,210 +1,0 @@
----
-title: "ssdtools"
-author: "Joe Thorley"
-date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
-vignette: >
- %\VignetteIndexEntry{ssdtools}
- %\VignetteEngine{knitr::rmarkdown}
- %\VignetteEncoding{UTF-8}
----
-
-```{r setup, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  fig.width = 3,
-  fig.height = 3
-)
-```
-
-
-## Introduction
-
-The `ssdtools` package fits and plots Species Sensitivity Distributions.
-This vignette describes how to use `ssdtools` in conjunction with 
-the `fitdistrplus` and `ggplot2` packages, which it installs and extends.
-
-## Inputting Data
-
-Data can be input from excel spreadsheets using `readxl::read_excel()` 
-and from csv files using `readr::read_csv()`.
-
-## Cullen Frey Plots
-
-The data can be visualized using a cullen frey plot of the skewness and kurtosis.
-Here we plot species sensitivity data for Boron.
-
-```{r, fig.width = 5, fig.height = 5}
-library(ssdtools)
-ssd_cfplot(boron_data)
-```
-
-## Fitting Distributions
-
-The function `ssd_fit_dists()` inputs a data frame (with the concentration values, 
-by default, in the `Conc` column) and fits one or more distributions.
-Here we fit the log-normal and weibull distributions to the Boron data.
-```{r}
-boron_data
-boron_dists2 <- ssd_fit_dists(boron_data, dists = c("lnorm", "weibull"))
-boron_dists2
-```
-
-The coefficients can be extracted using the `coef` function
-```{r}
-lapply(boron_dists2, coef)
-```
-
-## Model Diagnostics and Goodness of Fit
-
-The resultant `fitdists` object can be plotted to display model diagnostics plots for each fit.
-```{r, fig.width=6, fig.height=6, fig.show='hold'}
-plot(boron_dists2)
-```
-
-In addition, several goodness of fit statistics including the Akaike's Information Criteron corrected for sample size (`aicc`) can be generated using `ssd_gof()`.
-
-Moving forwards we consider the same six fits to Boron as Schwarz and Tillmanns (2017).
-```{r}
-boron_dists <- ssd_fit_dists(boron_data)
-ssd_gof(boron_dists)
-```
-For interpretation of the values see Schwarz and Tillmanns (2017).
-
-A subset of distributions from a `fitdists` can be selected using the subset function.
-```r
-subset(boron_dists, select = c("gamma", "weibull"))
-```
-
-## Plotting Fits
-
-The actual fits can be plotted using the ggplot2 generic `autoplot`.
-```{r, fig.width = 5 }
-library(ggplot2)
-autoplot(boron_dists)
-```
-
-As `autoplot` returns a `ggplot` object it can be modified prior to plotting, 
-which is extremely useful.
-```{r, fig.width = 5 }
-autoplot(boron_dists) + theme_bw() + ggtitle("Species Sensitivity Distributions for Boron")
-```
-
-## 5% Hazard Concentration
-
-Typically, species sensitivity distributions are fitted to estimate the 5% hazard concentration ($HC_5$).
-This is the concentration that affects just 5% of the species tested.
-The model-averaged $HC_5$ estimate (with 95% confidence limits) can be calculated as follows
-```{r}
-ssd_hc(boron_dists)
-```
-By default model averaging is based on the Akaike's Weights corrected for small sample size.
-To return the individual $HC_5$ estimates and their respective Akaike's Weights set `average = FALSE`
-```{r}
-ssd_hc(boron_dists, average = FALSE)
-```
[…]
