+++ NEWS.md
@@ -1,12 +1,10 @@
-- Removed dependencies: dplyr, magrittr, plyr, purrr
-- Moved from depends to imports: VGAM, fitdistrplus, graphics, ggplot, stats
-- Moved from imports to suggests: tibble
-
 # ssdtools 0.0.3
 
 - More information in DESCRIPTION
-- Added unused imported packages
 - Added CRAN badge
+- Removed dependencies: dplyr, magrittr, plyr, purrr
+- Moved from depends to imports: VGAM, fitdistrplus, graphics, ggplot, stats
+- Moved from imports to suggests: tibble
 
 # ssdtools 0.0.2
 
