+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: ssdtools
-Version: 0.0.3.9004
+Version: 0.0.3
 Title: Species Sensitivity Distributions
 Description: Species sensitivity distributions are cumulative probability 
   distributions which are fitted to toxicity concentrations for multiple species.

[…]
