+++ R/hc.R
@@ -31,8 +31,8 @@ ssd_hc <- function(x, ...) {
 #' @export
 #' @examples
 #' ssd_hc(boron_lnorm)
-ssd_hc.fitdist <- function(x, hc = 5L, nboot = 1001, level = 0.95, ...) {
-  check_vector(hc, c(1L,99L), length = 1)
+ssd_hc.fitdist <- function(x, hc = 5L, nboot = 1000, level = 0.95, ...) {
+  check_scalar(hc, c(1L,99L))
   predict(x, percent = hc, nboot = nboot, level = level)
 }
 
@@ -42,8 +42,8 @@ ssd_hc.fitdist <- function(x, hc = 5L, nboot = 1001, level = 0.95, ...) {
 #' \dontrun{
 #' ssd_hc(fluazinam_lnorm)
 #' }
-ssd_hc.fitdistcens <- function(x, hc = 5L, nboot = 1001, level = 0.95, ...) {
-  check_vector(hc, c(1L,99L), length = 1)
+ssd_hc.fitdistcens <- function(x, hc = 5L, nboot = 1000, level = 0.95, ...) {
+  check_scalar(hc, c(1L,99L)) 
   predict(x, percent = hc, nboot = nboot, level = level)
 }
 
@@ -53,8 +53,8 @@ ssd_hc.fitdistcens <- function(x, hc = 5L, nboot = 1001, level = 0.95, ...) {
 #' \dontrun{
 #' ssd_hc(boron_dists)
 #' }
-ssd_hc.fitdists <- function(x, hc = 5L, ic = "aicc", average = TRUE, nboot = 1001, level = 0.95, ...) {
-  check_vector(hc, c(1L,99L), length = 1)
+ssd_hc.fitdists <- function(x, hc = 5L, ic = "aicc", average = TRUE, nboot = 1000, level = 0.95, ...) {
+  check_scalar(hc, c(1L,99L))
   predict(x, percent = hc, ic = ic, nboot = nboot, average = average, level = level)
 }
 
@@ -64,7 +64,7 @@ ssd_hc.fitdists <- function(x, hc = 5L, ic = "aicc", average = TRUE, nboot = 100
 #' \dontrun{
 #' ssd_hc(fluazinam_dists)
 #' }
-ssd_hc.fitdistscens <- function(x, hc = 5L, ic = "aic", average = TRUE, nboot = 1001, level = 0.95, ...) {
-  check_vector(hc, c(1L,99L), length = 1)
+ssd_hc.fitdistscens <- function(x, hc = 5L, ic = "aic", average = TRUE, nboot = 1000, level = 0.95, ...) {
+  check_scalar(hc, c(1L,99L))
   predict(x, percent = hc, ic = ic, nboot = nboot, average = average, level = level)
 }
