+++ README.Rmd
@@ -66,8 +66,11 @@ ssd_gof(boron_dists)
 ```
 
 and the model-averaged 5% hazard concentration estimated using `ssd_hc`
-```{r, warning = FALSE}
-ssd_hc(boron_dists)
+```{r, eval = FALSE}
+boron_hc5 <- ssd_hc(boron_dists, nboot = 10000)
+```
+```{r}
+print(boron_hc5)
 ```
 
 Model-averaged predictions complete with confidence intervals can be produced using the `stats` generic `predict`
