+++ vignettes/ssdtools-manual.Rmd
@@ -23,13 +23,19 @@ knitr::opts_chunk$set(
 
 `ssdtools` is an R package to fit Species Sensitivity Distributions (SSDs) using Maximum Likelihood and model averaging.
 
-SSDs are cumulative probability distributions that are used to estimate the proportion of a range of species that are affected by a given concentration of a chemical.
+SSDs are cumulative probability distributions that are used to estimate the percent of species that are affected by a given concentration of a chemical.
 The concentration that affects 5% of the species is referred to as the 5% Hazard Concentration (HC).
 For more information on SSDs the reader is referred to @posthuma_species_2001.
 
 In order to use `ssdtools` you need to install R (see below) or use the Shiny [app](https://poissonconsulting.shinyapps.io/ssdtools/).
 The shiny app includes a user guide. 
 This vignette is a user manual for the R package.
+
+## Philosophy
+
+`ssdtools` provides the key functionality required to fit SSDs using Maximum Likelihood and model averaging in R.
+It is intended to be used in conjunction with other R packages such as `readr` to input data, `dplyr` to manipulate data and `ggplot2` to plot data.
+The general package design is in keeping with the tidyverse 
 
 ## Installing
 
@@ -49,6 +55,10 @@ library(ssdtools)
 To get additional information on a particular function just type `?` followed by the name of the function at the R console.
 For example `?ssd_gof` brings up the R documentation for the `ssdtools` goodness of fit function.
 
+For more information on using R the reader is referred to [R for Data Science](https://r4ds.had.co.nz) [@wickham_r_2016].
+
+If you discover a bug in `ssdtools` please file an issue with a [reprex](https://reprex.tidyverse.org/articles/reprex-dos-and-donts.html) (repeatable example) at <https://github.com/bcgov/ssdtools/issues>.
+
 ## Inputting Data
 
 Once the `ssdtools` package has been loaded the next task is to input some data.
@@ -65,86 +75,135 @@ data <- ssdtools::boron_data
 print(data)
 ```
 
+For cleaner printing of data frames install the `tibble` package
+```r
+install.packages("tibble")
+```
+
 ## Fitting Distributions
 
 The function `ssd_fit_dists()` inputs a data frame and fits one or more distributions.
-The user can specify a subset of the log-normal (`lnorm`), log-logistic (`llog`), log-Gumbel (`lgumbel`), Gompertz (`gompertz`), gamma (`gamma`) and Weibull (`weibull`) distributions and/or include the pareto (`pareto`) distribution using the `dists` argument.
+The user can specify a subset of the 
+
+- [log-normal](https://en.wikipedia.org/wiki/Log-normal_distribution) (`lnorm`),
+- [log-logistic](https://en.wikipedia.org/wiki/Log-logistic_distribution) (`llog`),
+- [log-Gumbel](https://en.wikipedia.org/wiki/Gumbel_distribution) (`lgumbel`),
+- [Gompertz](https://en.wikipedia.org/wiki/Gompertz_distribution) (`gompertz`),
+- [gamma](https://en.wikipedia.org/wiki/Gamma_distribution) (`gamma`) and 
+- [Weibull](https://en.wikipedia.org/wiki/Weibull_distribution) (`weibull`) 
+
+distributions and/or include the 
+
+- [Pareto](https://en.wikipedia.org/wiki/Pareto_distribution) (`pareto`)
+
+distribution using the `dists` argument.
 ```{r}
 dists <- ssd_fit_dists(data, dists = c("lnorm", "gompertz"))
 ```
 
 The coefficients can be extracted using the `coef` function.
-In and off themselves the coefficients are not that helpful.
+However, in and off themselves the coefficients are not that helpful.
 ```{r}
-lapply(dists, coef)
[…]

[…]
