### 2018-11-15 #1061 @nanhung
Overall, the package and paper look great. Just a few couples of questions and suggestions:

- For the title, I suggest removing the abbreviation SSDs.
- Does Carl Schwarz make any significant contributions to this paper or software?
- I find that there're only `ssdfit_dist` and `ssdfit_dists` rather than `ssdfit` in your package. Why not combine these two functions to one and named it ssdfit?
- I try to use `ggplot` with `geom_xribbon` and `geom_hcintercept` that you mentioned in summary. However, I don't know how to use `geom_xribbon`. Perhaps you can provide an example in the reference manual.
- Is it possible to use lower case "g" for `ggplot2` in reference?
- Does any of the references have doi? If so, please provide them?
- I can only find one reference in the reference manual. Is there some reference that you can add?
