+++ paper/paper.md
@@ -1,12 +1,12 @@
 ---
-title: 'ssdtools: An R package to fit Species Sensitivity Distributions (SSDs)'
+title: 'ssdtools: An R package to fit Species Sensitivity Distributions'
 authors:
 - affiliation: 1
   name: Joe Thorley
   orcid: 0000-0002-7683-4592
 - affiliation: 2
   name: Carl Schwarz
-date: '2018-11-01'
+date: '2018-11-15'
 output:
   pdf_document: default
   html_document:
@@ -39,8 +39,8 @@ The available Information Criteria are the Akaike Information Criterion (AIC), t
 Confidence intervals can be calculated for the fitted cumulative distribution function or specific hazard concentrations (percentiles).
 The confidence intervals are currently produced by parametric bootstrap resampling.
 
-`ssdtools` loads the `fitdistrplus` R package [@fitdistrplus] upon which it depends for model fitting, calculation of AIC and bootstrapping.
-`ssdtools` also loads the `ggplot2` R package [@ggplot2] which it extends by defining `ssdfit`, `xribbon` and `hcintersect` geometries to allow the user to produce custom SSD plots.
+`ssdtools` uses the `fitdistrplus` R package [@fitdistrplus] for model fitting, calculation of AIC and bootstrapping.
+`ssdtools` also extends the `ggplot2` R package [@ggplot2] by defining `ssdfit`, `xribbon` and `hcintersect` geometries to allow the user to produce custom SSD plots.
 
 # Acknowledgements
 
