+++ paper/paper.bib
@@ -20,11 +20,12 @@
 
   @Book{ggplot2,
     author = {Hadley Wickham},
-    title = {ggplot2: Elegant Graphics for Data Analysis},
+    title = {{g}gplot2: Elegant Graphics for Data Analysis},
     publisher = {Springer-Verlag New York},
     year = {2016},
     isbn = {978-3-319-24277-4},
-    url = {http://ggplot2.org}
+    url = {http://ggplot2.org},
+    doi = {10.1007/978-0-387-98141-3}
   }
 
  @Manual{r,
@@ -44,6 +45,6 @@
 	publisher = {Springer},
 	author = {Burnham, Kenneth P. and Anderson, David R.},
 	year = {2002},
-	keywords = {sgpd}
+	doi = {10.1007/b97636}
 }
 
