+++ README.md
@@ -148,7 +148,7 @@ If you would like to contribute to the package, please see our
 
 Please note that this project is released with a [Contributor Code of
 Conduct](CODE_OF_CONDUCT.md). By participating in this project you agree
-to abide by its terms
+to abide by its terms.
 
 ## License
 
