### 2018-11-16 #1082 @joethorley
Thanks @nanhung

My responses are as follows

- done
- yes he wrote the original R scripts on which ssdtools is based
- I decided to provide two functions so that the class of the returned object would not depend on the number of distributions used ie `ssd_fit_dist` returns a object of class `fitdist` while `ssd_fit_dists` returns a object of class `fitdists` which is a list of objects of class `fitdist`
- I've added a simple example of the use of each geom to the manual
- done
- I could only find a doi for two of the references (which I have added)
- The manual now has 5 references

The updated manual is at https://bcgov.github.io/ssdtools/articles/ssdtools-manual.html

Let me know if you require anymore information.
