### 2018-11-16 #1082 @nanhung
My concern is since you can use `ssd_fit_dists` to generate the same result as `ssd_fit_dist` such as,
```
data <- ssdtools::boron_data
ssd_fit_dist(data, dist = c("lnorm"))
ssd_fit_dists(data, dists = c("lnorm"))
```
Why users need to use `ssd_fit_dist` in their analysis?
