### 2018-11-16 #1082 @nanhung
Is it possible to use mutiple hc line in `ssd_plot`, such as,
```
ssd_plot(boron_data, boron_pred, color = "Group", label = "Species", hc = c(5L, 50L),
         xlab = "Concentration (mg/L)", ribbon = TRUE)
```
The current function only support 1 element.
