### 2018-11-16 f974a18
No longer export ssd_fit_dist() as ssd_fit_dists() renders redundant
