+++ NEWS.md
@@ -1,6 +1,7 @@
 - Added ssdtools-manual vignette
 - Changed predict() and ssd_hc() nboot argument from 1001 to 1000
-- Added hc5_boron data object.
+- Added hc5_boron data object
+- No longer export ssd_fit_dist() as ssd_fit_dists() renders redundant
 
 # ssdtools 0.0.3
 
