+++ NAMESPACE
@@ -58,7 +58,6 @@ export(rllog)
 export(rpareto)
 export(ssd_cfplot)
 export(ssd_ecd)
-export(ssd_fit_dist)
 export(ssd_fit_dists)
 export(ssd_gof)
 export(ssd_hc)
