### 2018-11-16 #1082 @joethorley
@nanhung `ssd_fit_dist()` is no longer exported to eliminate redundancy
