### 2018-11-16 dcdcf57
geom_hcintersect() now takes multiple values
