### 2018-11-16 #1082 @joethorley
@nanhung `geom_hcintersect()` now takes multiple arguments
