### 2018-11-16 #1082 @nanhung
Just a suggestion. Can you add an assigment (like "group_by") in `ssd_fit_dists` and `predict` that can simplify the approach if the user wants to apply the grouping method.

Here is an example code. I haven't use the `ssd_hc` in this case.

```
library(ssdtools)
library(tidyverse)
library(scales)

boron_dists_Fish <-  boron_data %>% filter(Group == "Fish") %>% ssd_fit_dists()
boron_dists_Amphibian <-  boron_data %>% filter(Group == "Amphibian") %>% ssd_fit_dists()
boron_dists_Invertebrate <- boron_data %>% filter(Group == "Invertebrate") %>% ssd_fit_dists()
boron_dists_Plant <-  boron_data %>% filter(Group == "Plant") %>% ssd_fit_dists()

boron_pred_Fish <- predict(boron_dists_Fish)
boron_pred_Amphibian <- predict(boron_dists_Amphibian)
boron_pred_Invertebrate <- predict(boron_dists_Invertebrate)
boron_pred_Plant <- predict(boron_dists_Plant)

boron_pred_Fish$Group <- "Fish"
boron_pred_Amphibian$Group <- "Amphibian"
boron_pred_Invertebrate$Group <- "Invertebrate"
boron_pred_Plant$Group <- "Plant"

x <- do.call(rbind, list(boron_pred_Fish, boron_pred_Amphibian, boron_pred_Invertebrate, boron_pred_Plant))

gp <- ggplot(x, aes_string(x = "est")) +
  geom_xribbon(aes_string(xmin = "lcl", xmax = "ucl", y = "percent/100"), alpha = 0.2) +
  geom_ssd(data = boron_data, aes_string(x = "Conc")) +
  facet_wrap(~Group) +
  scale_y_continuous("Species Affected (%)", labels = percent) +
  expand_limits(y = c(0, 1)) +
  xlab("Concentration (mg/L)") +
  coord_trans(x = "log10") +
  geom_line(aes_string(y = "percent/100")) +
  scale_x_continuous(breaks = scales::trans_breaks("log10", function(x) 10^x),
                     labels = comma_signif)

print(gp)
```
