+++ DESCRIPTION
@@ -42,7 +42,11 @@ Suggests:
     knitr,
     rmarkdown,
     testthat,
-    tibble
+    tibble,
+    tidyr,
+    readr,
+    purrr,
+    dplyr
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.1.1
