### 2018-11-16 #1082 @joethorley
@nanhung It's a frequent request. I've added an example to the manual demonstrating a more elegant approach to grouping ie

```
library(ssdtools)
library(tidyverse)

boron_datas <- nest(boron_data, -Group)
boron_datas <- mutate(boron_datas,
                      Fit = map(data, ssd_fit_dists, dists = "lnorm"),
                      Prediction = map(Fit, predict))
boron_datas <- unnest(boron_datas, Prediction)
```
