### 2018-11-16 #1082 @nanhung
How about the HC for each group? Can we use different HC values in this plot?
