### 2018-11-16 471eed5
add grouped hc5s
