### 2018-11-16 #1082 @joethorley
@nanhung I've also added to the manual a demonstration of how to plot HC5s for different groups

```
boron_hc5s <- filter(boron_datas, percent == 5)
gp %+% boron_datas +
  facet_wrap(~Group) +
  geom_hcintersect(data = boron_hc5s, aes(xintercept = est, yintercept = percent/100))
```
