### 2018-11-19 #1082 @nanhung
I'm trying to create the same plot in the example of `ssd_plot` by `geom_ssd` function. However, it seems that the `geom_ssd` doesn't support the different color by groups.

```
# Defind single color
ggplot(boron_data, aes(x = Conc)) +
  geom_ssd(colour = "red")

# Defind color by groups
ggplot(boron_data, aes(x = Conc)) +
  geom_ssd(colour = Group)
```
