### 2018-11-19 #1082 @joethorley
try
```
ggplot(boron_data, aes(x = Conc)) +
  geom_ssd(aes(colour = Group))
```
