### 2018-11-21 d54c4a8
Update ssdtools.R
Just adding ggplot2::
