### 2018-11-21 #1082 @ibarraespinosa
Hi @joethorley Interesting package. Demo is not running on line 42, consider adding ggplot2::autoplot(fits) or library(ggplto2) on ssdtools.R
