+++ demo/ssdtools.R
@@ -23,6 +23,7 @@
 #' devtools::install_github("bcgov/ssdtools")
 #' ```
 library(ssdtools)
+library(ggplot2)
 
 #' ### Data
 
