### 2018-11-21 #1082 @joethorley
@ibarraespinosa I accepted your pull request because it is more elegant than what I did
