+++ vignettes/references.bib
@@ -36,15 +36,17 @@
     url = {https://www.R-project.org/}
   }
 
-
 @book{burnham_model_2002,
-	address = {New York},
-	title = {Model {Selection} and {Multimodel} {Inference}: {A} {Practical} {Information}-{Theoretic} {Approach}},
-	volume = {Second Edition},
-	publisher = {Springer},
-	author = {Burnham, Kenneth P. and Anderson, David R.},
+	address = {New York, NY},
+	title = {Model {Selection} and {Multimodel} {Inference}},
+	isbn = {978-0-387-95364-9},
+	url = {http://link.springer.com/10.1007/b97636},
+	language = {en},
+	urldate = {2018-11-23},
+	publisher = {Springer New York},
+	editor = {Burnham, Kenneth P. and Anderson, David R.},
 	year = {2002},
-	keywords = {sgpd}
+	doi = {10.1007/b97636}
 }
 
 @techreport{schwarz_comparison_2017,
