+++ paper/paper.bib
@@ -37,14 +37,30 @@
     url = {https://www.R-project.org/}
   }
 
-
 @book{burnham_model_2002,
-	address = {New York},
-	title = {Model {Selection} and {Multimodel} {Inference}: {A} {Practical} {Information}-{Theoretic} {Approach}},
-	volume = {Second Edition},
-	publisher = {Springer},
-	author = {Burnham, Kenneth P. and Anderson, David R.},
+	address = {New York, NY},
+	title = {Model {Selection} and {Multimodel} {Inference}},
+	isbn = {978-0-387-95364-9},
+	url = {http://link.springer.com/10.1007/b97636},
+	language = {en},
+	urldate = {2018-11-23},
+	publisher = {Springer New York},
+	editor = {Burnham, Kenneth P. and Anderson, David R.},
 	year = {2002},
 	doi = {10.1007/b97636}
+}
+
+
+@book{millar_maximum_2011,
+	address = {Chichester, West Sussex},
+	series = {Statistics in practice},
+	title = {Maximum likelihood estimation and inference: with examples in {R}, {SAS}, and {ADMB}},
+	isbn = {978-0-470-09482-2},
+	shorttitle = {Maximum likelihood estimation and inference},
+	publisher = {Wiley},
+	author = {Millar, R. B.},
+	year = {2011},
+	note = {OCLC: ocn690090037},
+	keywords = {sgpd}
 }
 
