### 2018-11-23 477465d
update paper text
