+++ build.R
@@ -2,6 +2,8 @@ devtools::test()
 devtools::document()
 demo(ssdtools, ask = FALSE)
 # knitr::knit("README.Rmd")
-if(file.exists("DESCRIPTION")) unlink("docs", recursive = TRUE)
-pkgdown::build_site()
+if(FALSE) {
+  if(file.exists("DESCRIPTION")) unlink("docs", recursive = TRUE)
+  pkgdown::build_site()
+}
 devtools::check()
