+++ paper/paper.bib
@@ -39,13 +39,13 @@
 
 @book{burnham_model_2002,
 	address = {New York, NY},
-	title = {Model {Selection} and {Multimodel} {Inference}},
+	title = {Model {Selection} and {Multimodel} {Inference}: {A} {Practical} {Information-Theoretic} {Approach}},
 	isbn = {978-0-387-95364-9},
 	url = {http://link.springer.com/10.1007/b97636},
 	language = {en},
 	urldate = {2018-11-23},
 	publisher = {Springer New York},
-	editor = {Burnham, Kenneth P. and Anderson, David R.},
+	author = {Burnham, Kenneth P. and Anderson, David R.},
 	year = {2002},
 	doi = {10.1007/b97636}
 }
@@ -60,7 +60,5 @@
 	publisher = {Wiley},
 	author = {Millar, R. B.},
 	year = {2011},
-	note = {OCLC: ocn690090037},
-	keywords = {sgpd}
+	doi = {10.1002/9780470094846}
 }
-
