### 2018-11-23 ea06c62
remove print paper code
