### 2018-11-23 #1082 @joethorley
Thanks @ibarraespinosa!

I've added the missing doi, added the code for the plot and modified the text to make the problem and functionality clearer.
