### 2018-11-23 #1082 @ibarraespinosa
Hi @joethorley. Great, now, could you please add this doi:
 10.18637/jss.v064.i04

From https://www.jstatsoft.org/article/view/v064i04
