+++ paper/paper.bib
@@ -15,7 +15,8 @@
     volume = {64},
     number = {4},
     pages = {1--34},
-    url = {http://www.jstatsoft.org/v64/i04/}
+    url = {http://www.jstatsoft.org/v64/i04/},
+    doi = {10.18637/jss.v064.i04}
   }
 
   @Book{ggplot2,
