### 2018-11-25 962301c
update NEWS
