+++ NEWS.md
@@ -1,11 +1,10 @@
+# ssdtools 0.0.3
+
 - Added ssdtools-manual vignette
 - Changed predict() and ssd_hc() nboot argument from 1001 to 1000
 - Added hc5_boron data object
 - No longer export ssd_fit_dist() as ssd_fit_dists() renders redundant
 - geom_hcintersect() now takes multiple values
-
-# ssdtools 0.0.3
-
 - More information in DESCRIPTION
 - Added CRAN badge
 - Removed dependencies: dplyr, magrittr, plyr, purrr
