### 2018-11-28 #1082 @kyleniemeyer
Hi @joethorley, I just noticed that the affiliation info in the paper is slightly incomplete—could you extend it for both authors, e.g. with city and country info? (The way it would appear in a traditional article.)

Regarding the software DOI, it works ok here. The article DOI in the PDF will not be generated until the paper is actually published, though.
