### 2018-11-28 #1082 @joethorley
@kyleniemeyer I've fixed the affiliations but note that although the Review and Repository links work the Archive link does not - can you confirm this specific link works at your end as the archive DOI is already live (unlike the paper DOI). The full address is <https://doi.org/10.5281/zenodo.1651641>
