### 2018-11-28 #1082 @kyleniemeyer
@arfon when you get a chance, could you check out the PDF generation? The archive link is not working in the PDF—seems like something is going wrong.
