### 2018-11-29 #1082 @kyleniemeyer
@joethorley congrats on your paper acceptance into JOSS!

@ibarraespinosa & @nanhung thank you for your reviews, and thanks to @arfon for handling this submission.
